requirements.txt
streamlit
langchain-cohere
python-dotenv
pandas
numpy
pandasai
pandasai[langchain]
matplotlib
seaborn
chardet
tk 
tkintertable